re-structure tests directory
