+++ tests/commands/weather/current.test.js
@@ -18,21 +18,19 @@ jest.unstable_mockModule('ora', () => ({
 
 const buildWeatherLinkApiUrlMock = jest.fn();
 const checkForRequiredMock = jest.fn();
-const dateRangeIsValidMock = jest.fn();
 jest.unstable_mockModule('../../../lib/utils.js', () => {
   return {
     buildWeatherLinkApiUrl: buildWeatherLinkApiUrlMock,
     checkForRequired: checkForRequiredMock,
-    dateRangeIsValid: dateRangeIsValidMock,
   };
 });
 
-const { default: historic } = await import('../historic.js');
+const { default: current } = await import('../../../commands/weather/current.js');
 
 const logSpy = jest.spyOn(console, "log").mockImplementation(() => { });
 const dirSpy = jest.spyOn(console, "dir").mockImplementation(() => { });
 
-describe('wlbot weather historic', () => {
+describe('wlbot weather current', () => {
 
   beforeEach(() => {
     jest.resetModules();
@@ -60,14 +58,12 @@ describe('wlbot weather historic', () => {
   it('Success: Dry Run to Display Query URL', async () => {
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await historic(12345, 0, 1, { dryRun: true });
+    await current(12345, { dryRun: true, });
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith('http://unit.test');
@@ -80,15 +76,13 @@ describe('wlbot weather historic', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.resolve(mockSuccessfulApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(historic(12345, 0, 1, {})).resolves.toBe(mockSuccessfulApiCall);
+    await expect(current(12345, {})).resolves.toBe(mockSuccessfulApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledWith(mockSuccessfulApiCall.data, { depth: null });
     expect(logSpy).toHaveBeenCalledTimes(0);
@@ -101,57 +95,16 @@ describe('wlbot weather historic', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.resolve(mockSuccessfulApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(historic(12345, 0, 1, { raw: true })).resolves.toBe(mockSuccessfulApiCall);
+    await expect(current(12345, { raw: true })).resolves.toBe(mockSuccessfulApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(JSON.stringify(mockSuccessfulApiCall.data));
-    expect(oraFailMock).toHaveBeenCalledTimes(0);
-    expect(oraStartMock).toHaveBeenCalledTimes(0);
-    expect(oraSucceedMock).toHaveBeenCalledTimes(0);
-  });
-
-  it('Failure: Date Range is Invalid', () => {
-    buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
-    checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: false, msg: "Invalid datetime range." });
-
-    expect(historic(12345, -1, 0, {}));
-
-    expect(axiosGetMock).toHaveBeenCalledTimes(0);
-    expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
-    expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
-    expect(dirSpy).toHaveBeenCalledTimes(0);
-    expect(logSpy).toHaveBeenCalledTimes(1);
-    expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
-    expect(oraFailMock).toHaveBeenCalledTimes(1);
-    expect(oraFailMock).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
-    expect(oraStartMock).toHaveBeenCalledTimes(1);
-    expect(oraSucceedMock).toHaveBeenCalledTimes(0);
-  });
-
-  it('Failure: Date Range is Invalid (No Spinner)', () => {
-    buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
-    checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: false, msg: "Invalid datetime range." });
-
-    expect(historic(12345, -1, 0, { raw: true }));
-
-    expect(axiosGetMock).toHaveBeenCalledTimes(0);
-    expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
-    expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
-    expect(dirSpy).toHaveBeenCalledTimes(0);
-    expect(logSpy).toHaveBeenCalledTimes(1);
-    expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
     expect(oraFailMock).toHaveBeenCalledTimes(0);
     expect(oraStartMock).toHaveBeenCalledTimes(0);
     expect(oraSucceedMock).toHaveBeenCalledTimes(0);
@@ -161,15 +114,13 @@ describe('wlbot weather historic', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.reject(mockFailedApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(historic(12345, 0, 1, {})).rejects.toEqual(mockFailedApiCall);
+    await expect(current(12345, {})).rejects.toEqual(mockFailedApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Error'));
@@ -182,15 +133,13 @@ describe('wlbot weather historic', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.reject(mockFailedApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
-    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(historic(12345, 0, 1, { raw: true, })).rejects.toEqual(mockFailedApiCall);
+    await expect(current(12345, { raw: true, })).rejects.toEqual(mockFailedApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(JSON.stringify(mockFailedApiCall.response.data));
@@ -202,12 +151,11 @@ describe('wlbot weather historic', () => {
   it('Failure: Missing Environment Variables', async () => {
     checkForRequiredMock.mockReturnValue({ exist: false, missing: ['WEATHER_LINK_BASE_API_URL'] });
 
-    await historic(12345, 0, 1, {});
+    await current(12345, {});
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(0);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Missing Environment Variable'));
@@ -220,12 +168,11 @@ describe('wlbot weather historic', () => {
   it('Failure: Missing Environment Variables (No Spinner)', async () => {
     checkForRequiredMock.mockReturnValue({ exist: false, missing: ['WEATHER_LINK_BASE_API_URL'] });
 
-    await historic(12345, 0, 1, { dryRun: true, raw: true });
+    await current(12345, { dryRun: true, raw: true });
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
-    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(0);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Missing Environment Variable'));
