+++ tests/commands/weather/historic.test.js
@@ -18,19 +18,21 @@ jest.unstable_mockModule('ora', () => ({
 
 const buildWeatherLinkApiUrlMock = jest.fn();
 const checkForRequiredMock = jest.fn();
+const dateRangeIsValidMock = jest.fn();
 jest.unstable_mockModule('../../../lib/utils.js', () => {
   return {
     buildWeatherLinkApiUrl: buildWeatherLinkApiUrlMock,
     checkForRequired: checkForRequiredMock,
+    dateRangeIsValid: dateRangeIsValidMock,
   };
 });
 
-const { default: mine } = await import('../mine.js');
+const { default: historic } = await import('../../../commands/weather/historic.js');
 
 const logSpy = jest.spyOn(console, "log").mockImplementation(() => { });
 const dirSpy = jest.spyOn(console, "dir").mockImplementation(() => { });
 
-describe('wlbot metadata mine', () => {
+describe('wlbot weather historic', () => {
 
   beforeEach(() => {
     jest.resetModules();
@@ -42,18 +44,7 @@ describe('wlbot metadata mine', () => {
 
   const mockSuccessfulApiCall = {
     data: {
-      stations: [
-        {
-          station_id: 1,
-        },
-        {
-          station_id: 2,
-        },
-        {
-          station_id: 3,
-        },
-      ],
-      unit: "test",
+      unit: "test"
     },
   }
 
@@ -69,12 +60,14 @@ describe('wlbot metadata mine', () => {
   it('Success: Dry Run to Display Query URL', async () => {
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await mine({ dryRun: true, });
+    await historic(12345, 0, 1, { dryRun: true });
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith('http://unit.test');
@@ -87,15 +80,17 @@ describe('wlbot metadata mine', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.resolve(mockSuccessfulApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(mine({})).resolves.toStrictEqual([1, 2, 3]);
+    await expect(historic(12345, 0, 1, {})).resolves.toBe(mockSuccessfulApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(1);
-    expect(dirSpy).toHaveBeenCalledWith([1, 2, 3], { depth: null });
+    expect(dirSpy).toHaveBeenCalledWith(mockSuccessfulApiCall.data, { depth: null });
     expect(logSpy).toHaveBeenCalledTimes(0);
     expect(oraFailMock).toHaveBeenCalledTimes(0);
     expect(oraStartMock).toHaveBeenCalledTimes(1);
@@ -106,16 +101,57 @@ describe('wlbot metadata mine', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.resolve(mockSuccessfulApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(mine({ raw: true })).resolves.toBe(undefined);
+    await expect(historic(12345, 0, 1, { raw: true })).resolves.toBe(mockSuccessfulApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(JSON.stringify(mockSuccessfulApiCall.data));
+    expect(oraFailMock).toHaveBeenCalledTimes(0);
+    expect(oraStartMock).toHaveBeenCalledTimes(0);
+    expect(oraSucceedMock).toHaveBeenCalledTimes(0);
+  });
+
+  it('Failure: Date Range is Invalid', () => {
+    buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
+    checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: false, msg: "Invalid datetime range." });
+
+    expect(historic(12345, -1, 0, {}));
+
+    expect(axiosGetMock).toHaveBeenCalledTimes(0);
+    expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
+    expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
+    expect(dirSpy).toHaveBeenCalledTimes(0);
+    expect(logSpy).toHaveBeenCalledTimes(1);
+    expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
+    expect(oraFailMock).toHaveBeenCalledTimes(1);
+    expect(oraFailMock).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
+    expect(oraStartMock).toHaveBeenCalledTimes(1);
+    expect(oraSucceedMock).toHaveBeenCalledTimes(0);
+  });
+
+  it('Failure: Date Range is Invalid (No Spinner)', () => {
+    buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
+    checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: false, msg: "Invalid datetime range." });
+
+    expect(historic(12345, -1, 0, { raw: true }));
+
+    expect(axiosGetMock).toHaveBeenCalledTimes(0);
+    expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
+    expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
+    expect(dirSpy).toHaveBeenCalledTimes(0);
+    expect(logSpy).toHaveBeenCalledTimes(1);
+    expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Invalid datetime range.'));
     expect(oraFailMock).toHaveBeenCalledTimes(0);
     expect(oraStartMock).toHaveBeenCalledTimes(0);
     expect(oraSucceedMock).toHaveBeenCalledTimes(0);
@@ -125,13 +161,15 @@ describe('wlbot metadata mine', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.reject(mockFailedApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(mine({})).rejects.toEqual(mockFailedApiCall);
+    await expect(historic(12345, 0, 1, {})).rejects.toEqual(mockFailedApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Error'));
@@ -144,13 +182,15 @@ describe('wlbot metadata mine', () => {
     axiosGetMock.mockImplementationOnce(() => Promise.reject(mockFailedApiCall));
     buildWeatherLinkApiUrlMock.mockReturnValue('http://unit.test');
     checkForRequiredMock.mockReturnValue({ exist: true, missing: [] });
+    dateRangeIsValidMock.mockReturnValue({ isValid: true, msg: "" });
 
-    await expect(mine({ raw: true, })).rejects.toEqual(mockFailedApiCall);
+    await expect(historic(12345, 0, 1, { raw: true, })).rejects.toEqual(mockFailedApiCall);
 
     expect(axiosGetMock).toHaveBeenCalledTimes(1);
     expect(axiosGetMock).toHaveReturned();
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(1);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(1);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(JSON.stringify(mockFailedApiCall.response.data));
@@ -162,11 +202,12 @@ describe('wlbot metadata mine', () => {
   it('Failure: Missing Environment Variables', async () => {
     checkForRequiredMock.mockReturnValue({ exist: false, missing: ['WEATHER_LINK_BASE_API_URL'] });
 
-    await mine({});
+    await historic(12345, 0, 1, {});
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(0);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Missing Environment Variable'));
@@ -179,11 +220,12 @@ describe('wlbot metadata mine', () => {
   it('Failure: Missing Environment Variables (No Spinner)', async () => {
     checkForRequiredMock.mockReturnValue({ exist: false, missing: ['WEATHER_LINK_BASE_API_URL'] });
 
-    await mine({ dryRun: true, raw: true });
+    await historic(12345, 0, 1, { dryRun: true, raw: true });
 
     expect(axiosGetMock).toHaveBeenCalledTimes(0);
     expect(buildWeatherLinkApiUrlMock).toHaveBeenCalledTimes(0);
     expect(checkForRequiredMock).toHaveBeenCalledTimes(1);
+    expect(dateRangeIsValidMock).toHaveBeenCalledTimes(0);
     expect(dirSpy).toHaveBeenCalledTimes(0);
     expect(logSpy).toHaveBeenCalledTimes(1);
     expect(logSpy).toHaveBeenCalledWith(expect.stringContaining('Missing Environment Variable'));
